refactor(season): 季度重构

+++ core/src/main/kotlin/io/github/shoaky/sourcedownloader/api/ProcessorController.kt
@@ -70,7 +70,9 @@ private class ProcessorController(
     @GetMapping("/reload/{processorName}")
     fun reload(@PathVariable processorName: String) {
         val config = configOperator.getProcessorConfig(processorName)
-        processorManager.destroyProcessor(processorName)
+        if (processorManager.exists(processorName)) {
+            processorManager.destroyProcessor(processorName)
+        }
         processorManager.createProcessor(config)
     }
